+++ tests/database.js
@@ -126,7 +126,7 @@ test('database: get tags sorted by name descending', t => {
   )
 })
 
-test('database: get tags sorted by files', t => {
+test('database: get tags sorted by file count', t => {
   t.deepEqual(
     tags.get(1, 'files'),
     {
@@ -140,7 +140,7 @@ test('database: get tags sorted by files', t => {
   )
 })
 
-test('database: get tags sorted by files ascending', t => {
+test('database: get tags sorted by file count ascending', t => {
   t.deepEqual(
     tags.get(1, 'files', 'asc'),
     {
@@ -202,7 +202,7 @@ test('database: get tags containing sorted by name descending', t => {
   )
 })
 
-test('database: get tags containing sorted by files', t => {
+test('database: get tags containing sorted by file count', t => {
   t.deepEqual(
     tags.getContaining(1, 'lor', 'files'),
     {
@@ -214,7 +214,7 @@ test('database: get tags containing sorted by files', t => {
   )
 })
 
-test('database: get tags containing sorted by files ascending', t => {
+test('database: get tags containing sorted by file count ascending', t => {
   t.deepEqual(
     tags.getContaining(1, 'lor', 'files', 'asc'),
     {
@@ -299,6 +299,7 @@ test('database: get files', t => {
           size: 6672,
           width: 500,
           height: 500,
+          tagCount: 2,
           mediaUrl: originalBaseUrl +
             '/d2f5788f623cde1f0fb3dc801396fee235c67ed11d9452bfd765f1331587401d',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -310,6 +311,7 @@ test('database: get files', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -321,6 +323,7 @@ test('database: get files', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -332,6 +335,7 @@ test('database: get files', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -353,6 +357,7 @@ test('database: get files sorted ascending', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -364,6 +369,7 @@ test('database: get files sorted ascending', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -375,6 +381,7 @@ test('database: get files sorted ascending', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -386,6 +393,7 @@ test('database: get files sorted ascending', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -407,6 +415,7 @@ test('database: get files sorted by size', t => {
           size: 6672,
           width: 500,
           height: 500,
+          tagCount: 2,
           mediaUrl: originalBaseUrl +
             '/d2f5788f623cde1f0fb3dc801396fee235c67ed11d9452bfd765f1331587401d',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -418,6 +427,7 @@ test('database: get files sorted by size', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -429,6 +439,7 @@ test('database: get files sorted by size', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -440,6 +451,7 @@ test('database: get files sorted by size', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -461,6 +473,7 @@ test('database: get files sorted by size ascending', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -472,6 +485,7 @@ test('database: get files sorted by size ascending', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -483,6 +497,7 @@ test('database: get files sorted by size ascending', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -494,6 +509,7 @@ test('database: get files sorted by size ascending', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -515,6 +531,7 @@ test('database: get files sorted by width', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -526,6 +543,7 @@ test('database: get files sorted by width', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -537,6 +555,7 @@ test('database: get files sorted by width', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -548,6 +567,7 @@ test('database: get files sorted by width', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -569,6 +589,7 @@ test('database: get files sorted by width descending', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -580,6 +601,7 @@ test('database: get files sorted by width descending', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -591,6 +613,7 @@ test('database: get files sorted by width descending', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -602,6 +625,7 @@ test('database: get files sorted by width descending', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -623,6 +647,7 @@ test('database: get files sorted by height', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -634,6 +659,7 @@ test('database: get files sorted by height', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -645,6 +671,7 @@ test('database: get files sorted by height', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -656,6 +683,7 @@ test('database: get files sorted by height', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -677,6 +705,7 @@ test('database: get files sorted by height ascending', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -688,6 +717,7 @@ test('database: get files sorted by height ascending', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -699,6 +729,7 @@ test('database: get files sorted by height ascending', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -710,6 +741,7 @@ test('database: get files sorted by height ascending', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -731,6 +763,7 @@ test('database: get files sorted by mime', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -742,6 +775,7 @@ test('database: get files sorted by mime', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -753,6 +787,7 @@ test('database: get files sorted by mime', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -764,6 +799,7 @@ test('database: get files sorted by mime', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -785,6 +821,7 @@ test('database: get files sorted by mime descending', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -796,6 +833,7 @@ test('database: get files sorted by mime descending', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -807,6 +845,7 @@ test('database: get files sorted by mime descending', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -818,10 +857,127 @@ test('database: get files sorted by mime descending', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39'
+        }
+      ]
+    }
+  )
+})
+
+test('database: get files sorted by tag count', t => {
+  t.deepEqual(
+    files.get(1, 'tags'),
+    {
+      files: [
+        {
+          id: 1,
+          mime: 'image/png',
+          size: 5012,
+          width: 500,
+          height: 500,
+          tagCount: 6,
+          mediaUrl: originalBaseUrl +
+            '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+        },
+        {
+          id: 2,
+          mime: 'image/png',
+          size: 5779,
+          width: 500,
+          height: 500,
+          tagCount: 5,
+          mediaUrl: originalBaseUrl +
+            '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+        },
+        {
+          id: 3,
+          mime: 'image/png',
+          size: 6117,
+          width: 500,
+          height: 500,
+          tagCount: 4,
+          mediaUrl: originalBaseUrl +
+            '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+        },
+        {
+          id: 4,
+          mime: 'image/png',
+          size: 6665,
+          width: 500,
+          height: 500,
+          tagCount: 3,
+          mediaUrl: originalBaseUrl +
+            '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39'
+        }
+      ]
+    }
+  )
+})
+
+test('database: get files sorted by tag count ascending', t => {
+  t.deepEqual(
+    files.get(1, 'tags', 'asc'),
+    {
+      files: [
+        {
+          id: 5,
+          mime: 'image/png',
+          size: 6672,
+          width: 500,
+          height: 500,
+          tagCount: 2,
+          mediaUrl: originalBaseUrl +
+            '/d2f5788f623cde1f0fb3dc801396fee235c67ed11d9452bfd765f1331587401d',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/d2f5788f623cde1f0fb3dc801396fee235c67ed11d9452bfd765f1331587401d'
+        },
+        {
+          id: 4,
+          mime: 'image/png',
+          size: 6665,
+          width: 500,
+          height: 500,
+          tagCount: 3,
+          mediaUrl: originalBaseUrl +
+            '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39'
+        },
+        {
+          id: 3,
+          mime: 'image/png',
+          size: 6117,
+          width: 500,
+          height: 500,
+          tagCount: 4,
+          mediaUrl: originalBaseUrl +
+            '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+        },
+        {
+          id: 2,
+          mime: 'image/png',
+          size: 5779,
+          width: 500,
+          height: 500,
+          tagCount: 5,
+          mediaUrl: originalBaseUrl +
+            '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
         }
       ]
     }
@@ -839,6 +995,7 @@ test('database: get files sorted by namespace', t => {
           size: 6672,
           width: 500,
           height: 500,
+          tagCount: 2,
           mediaUrl: originalBaseUrl +
             '/d2f5788f623cde1f0fb3dc801396fee235c67ed11d9452bfd765f1331587401d',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -850,6 +1007,7 @@ test('database: get files sorted by namespace', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -861,6 +1019,7 @@ test('database: get files sorted by namespace', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -872,6 +1031,7 @@ test('database: get files sorted by namespace', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -893,6 +1053,7 @@ test('database: get files sorted by namespace descending', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -904,6 +1065,7 @@ test('database: get files sorted by namespace descending', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -915,6 +1077,7 @@ test('database: get files sorted by namespace descending', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -926,6 +1089,7 @@ test('database: get files sorted by namespace descending', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -947,6 +1111,7 @@ test('database: get files sorted by invalid namespace', t => {
           size: 6672,
           width: 500,
           height: 500,
+          tagCount: 2,
           mediaUrl: originalBaseUrl +
             '/d2f5788f623cde1f0fb3dc801396fee235c67ed11d9452bfd765f1331587401d',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -958,6 +1123,7 @@ test('database: get files sorted by invalid namespace', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -969,6 +1135,7 @@ test('database: get files sorted by invalid namespace', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -980,6 +1147,7 @@ test('database: get files sorted by invalid namespace', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1000,6 +1168,7 @@ test('database: get files by tags', t => {
         size: 5012,
         width: 500,
         height: 500,
+        tagCount: 6,
         mediaUrl: originalBaseUrl +
           '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
         thumbnailUrl: thumbnailsBaseUrl +
@@ -1020,6 +1189,7 @@ test('database: get files by excluded tags', t => {
           size: 6672,
           width: 500,
           height: 500,
+          tagCount: 2,
           mediaUrl: originalBaseUrl +
             '/d2f5788f623cde1f0fb3dc801396fee235c67ed11d9452bfd765f1331587401d',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1031,6 +1201,7 @@ test('database: get files by excluded tags', t => {
           size: 6665,
           width: 500,
           height: 500,
+          tagCount: 3,
           mediaUrl: originalBaseUrl +
             '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1042,6 +1213,7 @@ test('database: get files by excluded tags', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1062,6 +1234,7 @@ test('database: get files by tags and excluded tags', t => {
         size: 6117,
         width: 500,
         height: 500,
+        tagCount: 4,
         mediaUrl: originalBaseUrl +
           '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
         thumbnailUrl: thumbnailsBaseUrl +
@@ -1081,6 +1254,7 @@ test('database: get files by tags sorted ascending', t => {
         size: 5012,
         width: 500,
         height: 500,
+        tagCount: 6,
         mediaUrl: originalBaseUrl +
           '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
         thumbnailUrl: thumbnailsBaseUrl +
@@ -1101,6 +1275,7 @@ test('database: get files by tags sorted by size', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1112,6 +1287,7 @@ test('database: get files by tags sorted by size', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1123,6 +1299,7 @@ test('database: get files by tags sorted by size', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1144,6 +1321,7 @@ test('database: get files by tags sorted by size ascending', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1155,6 +1333,7 @@ test('database: get files by tags sorted by size ascending', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1166,6 +1345,7 @@ test('database: get files by tags sorted by size ascending', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1187,6 +1367,7 @@ test('database: get files by tags sorted by width', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1198,6 +1379,7 @@ test('database: get files by tags sorted by width', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1209,6 +1391,7 @@ test('database: get files by tags sorted by width', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1230,6 +1413,7 @@ test('database: get files by tags sorted by width ascending', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1241,6 +1425,7 @@ test('database: get files by tags sorted by width ascending', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1252,6 +1437,7 @@ test('database: get files by tags sorted by width ascending', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1273,6 +1459,7 @@ test('database: get files by tags sorted by height', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1284,6 +1471,7 @@ test('database: get files by tags sorted by height', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1295,6 +1483,7 @@ test('database: get files by tags sorted by height', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1316,6 +1505,7 @@ test('database: get files by tags sorted by height descending', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1327,6 +1517,7 @@ test('database: get files by tags sorted by height descending', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1338,6 +1529,7 @@ test('database: get files by tags sorted by height descending', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1359,6 +1551,7 @@ test('database: get files by tags sorted by mime', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1370,6 +1563,7 @@ test('database: get files by tags sorted by mime', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1381,6 +1575,7 @@ test('database: get files by tags sorted by mime', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1402,6 +1597,7 @@ test('database: get files by tags sorted by mime descending', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1413,6 +1609,7 @@ test('database: get files by tags sorted by mime descending', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1424,10 +1621,103 @@ test('database: get files by tags sorted by mime descending', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+        }
+      ]
+    }
+  )
+})
+
+test('database: get files by tags sorted by tag count', t => {
+  t.deepEqual(
+    files.getByTags(1, ['lorem', 'ipsum', 'dolor'], 'tags'),
+    {
+      files: [
+        {
+          id: 1,
+          mime: 'image/png',
+          size: 5012,
+          width: 500,
+          height: 500,
+          tagCount: 6,
+          mediaUrl: originalBaseUrl +
+            '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+        },
+        {
+          id: 2,
+          mime: 'image/png',
+          size: 5779,
+          width: 500,
+          height: 500,
+          tagCount: 5,
+          mediaUrl: originalBaseUrl +
+            '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+        },
+        {
+          id: 3,
+          mime: 'image/png',
+          size: 6117,
+          width: 500,
+          height: 500,
+          tagCount: 4,
+          mediaUrl: originalBaseUrl +
+            '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+        }
+      ]
+    }
+  )
+})
+
+test('database: get files by tags sorted by tag count ascending', t => {
+  t.deepEqual(
+    files.getByTags(1, ['lorem', 'ipsum', 'dolor'], 'tags', 'asc'),
+    {
+      files: [
+        {
+          id: 3,
+          mime: 'image/png',
+          size: 6117,
+          width: 500,
+          height: 500,
+          tagCount: 4,
+          mediaUrl: originalBaseUrl +
+            '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+        },
+        {
+          id: 2,
+          mime: 'image/png',
+          size: 5779,
+          width: 500,
+          height: 500,
+          tagCount: 5,
+          mediaUrl: originalBaseUrl +
+            '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+        },
+        {
+          id: 1,
+          mime: 'image/png',
+          size: 5012,
+          width: 500,
+          height: 500,
+          tagCount: 6,
+          mediaUrl: originalBaseUrl +
+            '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+          thumbnailUrl: thumbnailsBaseUrl +
+            '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
         }
       ]
     }
@@ -1447,6 +1737,7 @@ test('database: get files by tags sorted by namespace', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1458,6 +1749,7 @@ test('database: get files by tags sorted by namespace', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1469,6 +1761,7 @@ test('database: get files by tags sorted by namespace', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1492,6 +1785,7 @@ test('database: get files by tags sorted by namespace descending', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1503,6 +1797,7 @@ test('database: get files by tags sorted by namespace descending', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1514,6 +1809,7 @@ test('database: get files by tags sorted by namespace descending', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1537,6 +1833,7 @@ test('database: get files by tags sorted by invalid namespace', t => {
           size: 6117,
           width: 500,
           height: 500,
+          tagCount: 4,
           mediaUrl: originalBaseUrl +
             '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1548,6 +1845,7 @@ test('database: get files by tags sorted by invalid namespace', t => {
           size: 5779,
           width: 500,
           height: 500,
+          tagCount: 5,
           mediaUrl: originalBaseUrl +
             '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1559,6 +1857,7 @@ test('database: get files by tags sorted by invalid namespace', t => {
           size: 5012,
           width: 500,
           height: 500,
+          tagCount: 6,
           mediaUrl: originalBaseUrl +
             '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
           thumbnailUrl: thumbnailsBaseUrl +
@@ -1578,6 +1877,7 @@ test('database: get file by id', t => {
       size: 5012,
       width: 500,
       height: 500,
+      tagCount: 6,
       mediaUrl: originalBaseUrl +
         '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
       thumbnailUrl: thumbnailsBaseUrl +